+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# 0.3.1
+
+* Bug - Log directory is now automatically created when the helper runs
+
+# 0.3.0
+
+* Feature - Added support for PrivateLink endpoints
+
 # 0.2.0
 
 * Feature - Added support for FIPS endpoints